fix converstion errors

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: rpassafa <rpassafa@student.42.us>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/10 15:29:27 by rpassafa          #+#    #+#             */
-/*   Updated: 2016/11/16 11:38:58 by rpassafa         ###   ########.us       */
+/*   Updated: 2016/11/17 15:38:59 by rpassafa         ###   ########.us       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,12 +51,12 @@ void	setflags(s_flags **flag);
 void	setsymb(s_flags **flag, int *findex, const char *format);
 void	findflags(s_flags **flag, int *findex, const char *format);
 //extra functions
-int findsize(int value, int base);
+int findsize(intmax_t value, int base);
 char *betterjoin(char *first, char *second);
 //handle
 uintmax_t		ft_pow(uintmax_t nb, int pow);
-char *ft_itoa_base(int value, int base);
-char *ft_itoa_baselow(int value, int base);
+char *ft_itoa_base(intmax_t value, int base);
+char *ft_itoa_baselow(intmax_t value, int base);
 //flag handlers
 int checkzeroflag(s_flags **flag);
 int findprecision(const char *format, int *findex);
@@ -64,7 +64,7 @@ int checkoptions(s_flags **flag);
 //handle options
 int findprecision(const char *format, int *findex);
 char *flagformating(char *str, s_flags **flag);
-void isnegative(s_flags **flag, int data);
+void isnegative(s_flags **flag, intmax_t data);
 //printf
 int ft_printf(const char *format, ...);
 #endif
